requirements.txt
pandas  
matplotlib
Pillow
python-docx